feat(scss): add default.scss file with global styles
The default.scss file is added to the scss folder. It imports the Roboto font from Google Fonts and the colors.scss file. It sets the global styles for the HTML and body elements, including margin, padding, box-sizing, font-size, background-color, and font-family. This file will be used to apply consistent styles across the application.